tools/importer-rest-api-specs: refactoring each of the field processors out
