+++ tools/importer-rest-api-specs/components/schema/processors/fields.go
@@ -1,11 +1,6 @@
 package processors
 
 import (
-	"fmt"
-	"regexp"
-	"strings"
-
-	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/parser/cleanup"
 	"github.com/hashicorp/pandora/tools/sdk/resourcemanager"
 )
 
@@ -25,85 +20,6 @@ var NamingRules = []FieldNameProcessor{
 	fieldNamePluralToSingular{},
 	fieldNameRemoveResourcePrefix{},
 	fieldNameRenameBoolean{},
-}
-
-type fieldNameExists struct{}
-
-func (fieldNameExists) ProcessField(fieldName string, metadata FieldMetadata) (*string, error) {
-	_, ok := metadata.Model.Fields[fieldName]
-	if !ok {
-		return nil, fmt.Errorf("%s was not found in %+v", fieldName, metadata.Model.Fields)
-	}
-	return nil, nil
-}
-
-type fieldNameRemoveIsPrefix struct{}
-
-func (fieldNameRemoveIsPrefix) ProcessField(fieldName string, metadata FieldMetadata) (*string, error) {
-	re := regexp.MustCompile("^Is[A-Z][a-z]*")
-	if re.MatchString(fieldName) {
-		updatedName := fieldName[2:]
-		return &updatedName, nil
-	}
-	return nil, nil
-}
-
-type fieldNamePluralToSingular struct{}
-
-func (fieldNamePluralToSingular) ProcessField(fieldName string, metadata FieldMetadata) (*string, error) {
-	if metadata.Model.Fields[fieldName].ObjectDefinition.Type == resourcemanager.ListApiObjectDefinitionType {
-		updatedName := cleanup.GetSingular(fieldName)
-		return &updatedName, nil
-	}
-	return nil, nil
-}
-
-type fieldNameRenameBoolean struct{}
-
-func (fieldNameRenameBoolean) ProcessField(fieldName string, metadata FieldMetadata) (*string, error) {
-	if metadata.Model.Fields[fieldName].ObjectDefinition.Type == resourcemanager.BooleanApiObjectDefinitionType {
-		var updatedFieldName *string
-
-		if strings.HasPrefix(strings.ToLower(fieldName), "enabled") {
-			updated := fmt.Sprintf("%sEnabled", fieldName[7:])
-			return &updated, nil
-		}
-		if strings.HasPrefix(strings.ToLower(fieldName), "disabled") {
-			updated := fmt.Sprintf("%sDisabled", fieldName[7:])
-			return &updated, nil
-		}
-		// flip `enable_X` / `disable_X` prefix
-		if strings.HasPrefix(strings.ToLower(fieldName), "enable") {
-			updated := fmt.Sprintf("%sEnabled", fieldName[6:])
-			updatedFieldName = &updated
-		}
-		if strings.HasPrefix(strings.ToLower(fieldName), "disable") {
-			updated := fmt.Sprintf("%sDisabled", fieldName[7:])
-			updatedFieldName = &updated
-		}
-		// change `allow_X` prefix -> `X_enabled`
-		if strings.HasPrefix(strings.ToLower(fieldName), "allow") {
-			updated := fmt.Sprintf("%sEnabled", fieldName[5:])
-			updatedFieldName = &updated
-		}
-		// change `allowed_X` prefix -> `X_enabled`
-		if strings.HasPrefix(strings.ToLower(fieldName), "allowed") {
-			updated := fmt.Sprintf("%sEnabled", fieldName[7:])
-			updatedFieldName = &updated
-		}
-		return updatedFieldName, nil
-	}
-	return nil, nil
-}
-
-type fieldNameRemoveResourcePrefix struct{}
-
-func (fieldNameRemoveResourcePrefix) ProcessField(fieldName string, metadata FieldMetadata) (*string, error) {
-	if strings.HasPrefix(fieldName, metadata.TerraformDetails.ResourceName) {
-		updatedName := strings.Replace(fieldName, metadata.TerraformDetails.ResourceName, "", 1)
-		return &updatedName, nil
-	}
-	return nil, nil
 }
 
 //TODO: if it's a List[Reference] and the model contains a single field `Id` then flatten this into `_ids`.
